+++ Runtime/ToastPosition.cs
@@ -1,0 +1,10 @@
+namespace OneM.ToastSystem
+{
+    public enum ToastPosition
+    {
+        Top,
+        Bottom,
+        Left,
+        Right
+    }
+}