+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ### Added
+- Option to select Toast Position
 - Serialized Dictionary package dependency
 
 ## [1.2.0] - 2026-03-28